requirements.txt
discord
asyncio
PyNaCl
ffmpeg-python
discord.py
python-dotenv
